+++ src/th06_102h.exe/Global/FUN_0040dbc0.c
@@ -28,7 +28,7 @@ void FUN_0040dbc0(int param_1,int param_2)
       local_64[1] = (float)(fVar1 * (float10)*(float *)(param_1 + 0x9c0)) + 192.0;
       local_64[2] = (float)(fVar5 * (float10)*(float *)(param_1 + 0x9c0)) + 224.0;
       local_54 = local_68 + *(float *)(param_1 + 0x9b8);
-      FUN_004145c0(local_64);
+      FUN_004145c0((int)local_64);
       local_68 = 6.283185 / (float)iVar2 + local_68;
     }
   }
